Refactor instrument table state and helpers
